requirements.txt
backports.entry-points-selectable==1.1.1
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
cycler==0.10.0
distlib==0.3.3
docopt==0.6.2
filelock==3.4.0
Flask==2.0.1
Flask-Dance==5.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.1.1
platformdirs==2.4.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.25
threadpoolctl==3.0.0
tqdm==4.62.3
urllib3==1.26.7
URLObject==2.4.3
virtualenv==20.10.0
Werkzeug==2.0.1
yarg==0.1.9
